update to use dirty
