+++ src/app/Scenes/SellWithArtsy/ConsignmentInquiry/ConsignmentInquiryScreen.tsx
@@ -1,6 +1,5 @@
 import { ActionType, ContextModule, OwnerType } from "@artsy/cohesion"
 import { SentConsignmentInquiry } from "@artsy/cohesion/dist/Schema/Events/Consignments"
-import { useNavigation } from "@react-navigation/native"
 import { ConsignmentInquiryScreenMutation } from "__generated__/ConsignmentInquiryScreenMutation.graphql"
 import { FancyModal } from "app/Components/FancyModal/FancyModal"
 import { FancyModalHeader } from "app/Components/FancyModal/FancyModalHeader"
@@ -8,7 +7,7 @@ import { useToast } from "app/Components/Toast/toastHook"
 import { defaultEnvironment } from "app/relay/createEnvironment"
 import { FormikProvider, useFormik } from "formik"
 import { CloseIcon } from "palette"
-import { useEffect, useState } from "react"
+import { useState } from "react"
 import { useTracking } from "react-tracking"
 import { commitMutation, Environment, graphql } from "relay-runtime"
 import { ArtsyKeyboardAvoidingView } from "shared/utils"
@@ -78,28 +77,10 @@ export const ConsignmentInquiryScreen: React.FC<InquiryScreenProps> = ({
   phone,
   userId,
 }) => {
-  const [formIsDirty, setFormIsDirty] = useState(false)
   const [showAbandonModal, setShowAbandonModal] = useState(false)
   const [showConfirmedModal, setShowConfirmedModal] = useState(false)
-  const navigation = useNavigation()
   const { trackEvent } = useTracking()
   const toast = useToast()
-
-  useEffect(() => {
-    const backListener = navigation.addListener("beforeRemove", (e) => {
-      e.preventDefault()
-      if (showAbandonModal || showConfirmedModal) {
-        navigation.dispatch(e.data.action)
-        return
-      }
-      if (formIsDirty) {
-        setShowAbandonModal(true)
-        return
-      }
-      navigation.dispatch(e.data.action)
-    })
-    return backListener
-  }, [navigation, showAbandonModal, showConfirmedModal, formIsDirty])
 
   const formik = useFormik<InquiryFormikSchema>({
     validateOnChange: true,
@@ -139,18 +120,16 @@ export const ConsignmentInquiryScreen: React.FC<InquiryScreenProps> = ({
     validationSchema: ValidationSchema,
   })
 
-  const updateDirty = () => {
-    if (formIsDirty) {
-      return
-    }
-    setFormIsDirty(true)
-  }
+  const canPopScreen = showAbandonModal || showConfirmedModal
 
   return (
     <FormikProvider value={formik}>
       <>
         <ArtsyKeyboardAvoidingView>
-          <ConsignmentInquiryForm onAnyFieldEdited={updateDirty} />
+          <ConsignmentInquiryForm
+            confirmLeaveEdit={(v) => setShowAbandonModal(v)}
+            canPopScreen={canPopScreen}
+          />
         </ArtsyKeyboardAvoidingView>
 
         <FancyModal visible={showAbandonModal && !showConfirmedModal}>
